backend/requirements.txt
flask==2.3.2
opencv-python==4.7.0.72
numpy==1.26.4
scipy==1.11.3
scikit-learn==1.3.0
xgboost==1.7.6
pillow==10.0.0
torch==2.0.1
